Moved PermissionsChecker to permissions module
