+++ collect_app/build.gradle
@@ -250,6 +250,7 @@ dependencies {
     implementation project(':location')
     implementation project(':geo')
     implementation project(':errors')
+    implementation project(':permissions')
 
     implementation Dependencies.androidx_startup
     implementation Dependencies.androidx_legacy_support
