+++ collect_app/src/main/java/org/odk/collect/android/formentry/BackgroundAudioViewModel.java
@@ -17,12 +17,12 @@ import org.odk.collect.android.analytics.AnalyticsUtils;
 import org.odk.collect.android.formentry.audit.AuditEvent;
 import org.odk.collect.android.formentry.audit.AuditEventLogger;
 import org.odk.collect.android.javarosawrapper.FormController;
-import org.odk.collect.androidshared.system.PermissionsChecker;
 import org.odk.collect.androidshared.livedata.MutableNonNullLiveData;
 import org.odk.collect.androidshared.livedata.NonNullLiveData;
 import org.odk.collect.audiorecorder.recorder.Output;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.audiorecorder.recording.RecordingSession;
+import org.odk.collect.permissions.PermissionsChecker;
 import org.odk.collect.shared.Settings;
 import org.odk.collect.utilities.Clock;
