+++ collect_app/src/test/java/org/odk/collect/android/audio/AudioRecordingControllerFragmentTest.java
@@ -25,7 +25,6 @@ import org.odk.collect.android.R;
 import org.odk.collect.android.formentry.BackgroundAudioViewModel;
 import org.odk.collect.android.formentry.FormEntryViewModel;
 import org.odk.collect.android.injection.config.AppDependencyModule;
-import org.odk.collect.androidshared.system.PermissionsChecker;
 import org.odk.collect.android.preferences.source.SettingsProvider;
 import org.odk.collect.android.support.CollectHelpers;
 import org.odk.collect.android.utilities.ExternalWebPageHelper;
@@ -34,6 +33,7 @@ import org.odk.collect.audiorecorder.recorder.Output;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.audiorecorder.testsupport.StubAudioRecorder;
 import org.odk.collect.fragmentstest.DialogFragmentTest;
+import org.odk.collect.permissions.PermissionsChecker;
 import org.odk.collect.utilities.Clock;
 import org.robolectric.annotation.Config;
