+++ collect_app/src/test/java/org/odk/collect/android/formentry/BackgroundAudioViewModelTest.java
@@ -12,9 +12,9 @@ import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.formentry.audit.AuditEvent;
 import org.odk.collect.android.formentry.audit.AuditEventLogger;
 import org.odk.collect.android.javarosawrapper.FormController;
-import org.odk.collect.androidshared.system.PermissionsChecker;
 import org.odk.collect.audiorecorder.recorder.Output;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
+import org.odk.collect.permissions.PermissionsChecker;
 import org.odk.collect.shared.Settings;
 import org.odk.collect.testshared.RobolectricHelpers;
 import org.odk.collect.utilities.Clock;
