+++ collect_app/src/test/java/org/odk/collect/android/projects/QrCodeProjectCreatorDialogTest.kt
@@ -32,7 +32,7 @@ import org.odk.collect.android.permissions.PermissionsProvider
 import org.odk.collect.android.support.CollectHelpers
 import org.odk.collect.android.utilities.CodeCaptureManagerFactory
 import org.odk.collect.android.views.BarcodeViewDecoder
-import org.odk.collect.androidshared.system.PermissionsChecker
+import org.odk.collect.permissions.PermissionsChecker
 import org.odk.collect.projects.ProjectsRepository
 import org.odk.collect.testshared.RobolectricHelpers
 import org.robolectric.shadows.ShadowToast
