+++ permissions/build.gradle
@@ -36,5 +36,5 @@ android {
 }
 
 dependencies {
-
+    implementation Dependencies.androidx_core_ktx
 }
